+++ e2e/src/app.po.ts
@@ -1,11 +1,15 @@
-import { browser, by, element } from 'protractor';
+import {browser, by, element} from 'protractor';
 
 export class AppPage {
-  navigateTo() {
-    return browser.get(browser.baseUrl) as Promise<any>;
+  navigateTo(path: string) {
+    return browser.get(path) as Promise<any>;
   }
 
-  getTitleText() {
-    return element(by.css('app-root .content span')).getText() as Promise<string>;
+  getElementTextByTagName(tag: string) {
+    return element(by.tagName(tag)).getText() as Promise<string>;
+  }
+
+  getElementAttributeByCss(css: string, attribute: string) {
+    return element(by.css(css)).getAttribute(attribute) as Promise<string>;
   }
 }