+++ src/app/components/home-minifier/home-minifier.component.ts
@@ -10,6 +10,7 @@ import {LanguagesEnum} from '../../shared/enums/Languages.enum';
 import {MinifyAllClass} from '@josee9988/minifyall';
 import * as Terser from "terser";
 import {Meta, Title} from "@angular/platform-browser";
+import {environment} from "../../../environments/environment";
 
 @Component({
   selector: 'app-home-component',
@@ -20,7 +21,8 @@ export class HomeMinifier extends Forms implements OnInit {
   languageSelected = 0;
   minifiedCode = '';
   isProcessing = false;
-  isHexMinifierEnabled: boolean;
+  domainUrl: string = environment.domainUrl;
+  isHexMinifierEnabled = false;
   nonMinifiedCode: FormControl;
   languages: ILanguagesInterface[] = [
     {value: 0, viewValue: 'Auto detect on paste', faIcon: 'fas fa-magic'},
@@ -135,5 +137,9 @@ export class HomeMinifier extends Forms implements OnInit {
           'Unexpected error while minifying. Please tell us how this happened!', SnackbarTypeEnum.error);
         break;
     }
+  }
+
+  test() {
+    console.log(this.isHexMinifierEnabled)
   }
 }